move auth provider logic into providerService
