+++ app/services/authService/index.ts
@@ -1,1 +1,2 @@
 export * from './_amplify';
+export { useAuthSession } from '../providerService/authProvider';
