+++ app/services/providerService/index.ts
@@ -4,4 +4,5 @@
  * ----------------------------------------------- */
 
 export { default as AppRootProvider } from './_appRootProvider';
+export { AuthProvider, useAuthSession } from './authProvider';
 export { default as ToastProvider } from './_toastProvider';
